+++ src/Analyser/StaticAnalyser.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace Mihaeu\PhpDependencies\Analyser;
 
 use Mihaeu\PhpDependencies\Dependencies\DependencyMap;
+use Mihaeu\PhpDependencies\Exceptions\ParserException;
 use Mihaeu\PhpDependencies\OS\PhpFile;
 use Mihaeu\PhpDependencies\OS\PhpFileSet;
+use PhpParser\Error;
 use PhpParser\NodeTraverser;
 use PhpParser\NodeVisitor\NameResolver;
 
@@ -42,9 +44,13 @@ class StaticAnalyser
     public function analyse(PhpFileSet $files) : DependencyMap
     {
         $files->each(function (PhpFile $file) {
-            $this->nodeTraverser->traverse(
-                $this->parser->parse($file->code())
-            );
+            try {
+                $this->nodeTraverser->traverse(
+                    $this->parser->parse($file->code())
+                );
+            } catch (Error $e) {
+                throw new ParserException($e->getMessage(), $file->file()->getPathname());
+            }
         });
 
         return $this->dependencyInspectionVisitor->dependencies();