+++ tests/unit/Analyser/StaticAnalyserTest.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace Mihaeu\PhpDependencies\Analyser;
 
 use Mihaeu\PhpDependencies\Dependencies\DependencyMap;
+use Mihaeu\PhpDependencies\Exceptions\ParserException;
 use Mihaeu\PhpDependencies\OS\PhpFile;
 use Mihaeu\PhpDependencies\OS\PhpFileSet;
+use PhpParser\Error;
 use PhpParser\NodeTraverser;
 use PHPUnit\Framework\TestCase;
 use PHPUnit_Framework_MockObject_MockObject;
@@ -46,5 +48,17 @@ class StaticAnalyserTest extends TestCase
         $phpFile->method('code')->willReturn('');
         $dependencies = $this->analyser->analyse((new PhpFileSet())->add($phpFile));
         assertEquals(new DependencyMap(), $dependencies);
+    }
+
+    public function testEnrichesExceptionWhenParserThrows(): void
+    {
+        $phpFile = $this->createMock(PhpFile::class);
+        $phpFile->method('code')->willReturn('');
+        $phpFile->method('file')->willReturn(new \SplFileInfo('test.php'));
+        $this->parser->method('parse')
+            ->willThrowException(new Error('', []));
+
+        $this->expectException(ParserException::class);
+        $this->analyser->analyse((new PhpFileSet())->add($phpFile));
     }
 }